Read Formatting of All Sheets in an Excel Workbook

Retrieves all sheets of a Microsoft Excel workbook and identifies the formatting of each value (including column headers and blank cells).

read_xl_format(file_name = NULL)

Arguments

file_name

(character) Name of (and path to) the Excel workbook

Value

(data frame) One row per cell in the dataframe with a column for each type of relevant formatting and its 'address' within the original Excel workbook

Examples

# Identify the formatting of every cell in all sheets of an Excel file
read_xl_format(file_name = system.file("extdata", "faux_data.xlsx", package = "scicomptools"))
#>      sheet address row col cell_contents               comment    formula  bold
#> 1  my_data      A1   1   1         col_1                  <NA>       <NA>  TRUE
#> 2  my_data      B1   1   2         col_2                  <NA>       <NA>  TRUE
#> 3  my_data      C1   1   3         col_3                  <NA>       <NA>  TRUE
#> 4  my_data      D1   1   4         col_4                  <NA>       <NA>  TRUE
#> 5  my_data      E1   1   5         col_5 Nick Lyon:Hello world       <NA>  TRUE
#> 6  my_data      A2   2   1             a                  <NA>       <NA>  TRUE
#> 7  my_data      B2   2   2             A                  <NA>       <NA>  TRUE
#> 8  my_data      C2   2   3             2                  <NA>       <NA>  TRUE
#> 9  my_data      D2   2   4           234                  <NA>       <NA>  TRUE
#> 10 my_data      E2   2   5           236                  <NA> SUM(C2:D2)  TRUE
#> 11 my_data      A3   3   1             b                  <NA>       <NA>  TRUE
#> 12 my_data      B3   3   2             A                  <NA>       <NA>  TRUE
#> 13 my_data      C3   3   3            43                  <NA>       <NA>  TRUE
#> 14 my_data      D3   3   4           435                  <NA>       <NA>  TRUE
#> 15 my_data      E3   3   5           478                  <NA> SUM(C3:D3)  TRUE
#> 16 my_data      A4   4   1             c                  <NA>       <NA>  TRUE
#> 17 my_data      B4   4   2             A                  <NA>       <NA>  TRUE
#> 18 my_data      C4   4   3             2                  <NA>       <NA>  TRUE
#> 19 my_data      D4   4   4            32                  <NA>       <NA>  TRUE
#> 20 my_data      E4   4   5            34                  <NA> SUM(C4:D4)  TRUE
#> 21 my_data      A5   5   1             d                  <NA>       <NA>  TRUE
#> 22 my_data      B5   5   2             B                  <NA>       <NA>  TRUE
#> 23 my_data      C5   5   3             2                  <NA>       <NA>  TRUE
#> 24 my_data      D5   5   4            45                  <NA>       <NA>  TRUE
#> 25 my_data      E5   5   5            47                  <NA> SUM(C5:D5)  TRUE
#> 26 my_data      A6   6   1             e                  <NA>       <NA>  TRUE
#> 27 my_data      B6   6   2             B                  <NA>       <NA>  TRUE
#> 28 my_data      C6   6   3             2                  <NA>       <NA>  TRUE
#> 29 my_data      D6   6   4             4                  <NA>       <NA>  TRUE
#> 30 my_data      E6   6   5             6                  <NA> SUM(C6:D6)  TRUE
#> 31  data_2      A1   1   1         col_2                  <NA>       <NA> FALSE
#> 32  data_2      B1   1   2   temperature                  <NA>       <NA> FALSE
#> 33  data_2      A2   2   1             A                  <NA>       <NA> FALSE
#> 34  data_2      B2   2   2            10                  <NA>       <NA> FALSE
#> 35  data_2      A3   3   1             B                  <NA>       <NA> FALSE
#> 36  data_2      B3   3   2            20                  <NA>       <NA>  TRUE
#>    italic underline font_size font_color cell_color
#> 1   FALSE      <NA>        12   FF000000       <NA>
#> 2   FALSE      <NA>        12   FF000000       <NA>
#> 3   FALSE      <NA>        12   FF000000       <NA>
#> 4   FALSE      <NA>        12   FF000000       <NA>
#> 5   FALSE      <NA>        12   FF000000       <NA>
#> 6   FALSE      <NA>        12   FF000000       <NA>
#> 7   FALSE      <NA>        12   FF000000       <NA>
#> 8   FALSE      <NA>        12   FF000000   FFFFFFFF
#> 9   FALSE      <NA>        12   FF000000       <NA>
#> 10   TRUE      <NA>        12   FF000000       <NA>
#> 11  FALSE      <NA>        12   FF000000       <NA>
#> 12  FALSE      <NA>        12   FF000000       <NA>
#> 13  FALSE      <NA>        12   FF000000   FFFFFFFF
#> 14  FALSE      <NA>        12   FF000000       <NA>
#> 15   TRUE      <NA>        12   FF000000       <NA>
#> 16  FALSE      <NA>        12   FF000000       <NA>
#> 17  FALSE      <NA>        12   FF000000       <NA>
#> 18  FALSE      <NA>        12   FF000000       <NA>
#> 19  FALSE    single        12   FF000000       <NA>
#> 20   TRUE      <NA>        12   FF000000       <NA>
#> 21  FALSE      <NA>        12   FF000000       <NA>
#> 22  FALSE      <NA>        12   FF000000       <NA>
#> 23  FALSE      <NA>        12   FF000000       <NA>
#> 24  FALSE    single        12   FF000000       <NA>
#> 25  FALSE      <NA>        12   FF000000       <NA>
#> 26  FALSE      <NA>        12   FF000000       <NA>
#> 27  FALSE      <NA>        12   FF000000       <NA>
#> 28  FALSE      <NA>        12   FF000000       <NA>
#> 29  FALSE      <NA>        12   FF000000       <NA>
#> 30  FALSE      <NA>        12   FF000000       <NA>
#> 31  FALSE      <NA>        12   FF000000       <NA>
#> 32  FALSE      <NA>        12   FFFF0000       <NA>
#> 33   TRUE      <NA>        12   FF000000       <NA>
#> 34  FALSE      <NA>        12   FF000000       <NA>
#> 35  FALSE    single        12   FF000000       <NA>
#> 36  FALSE      <NA>        12   FF000000       <NA>